adding the function

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -10,7 +10,7 @@
 
 void print_rev(char *s)
 {
-	int length = strlen(s);
+	size_t length = strlen(s);
 	
 	int i;
